Limit absolute times to reasonable daylight hours
In the case of EX raids, catch the case where an unqualified hour
is used as the start time (e.g., 1:00) and the start time is set
for the middle of the night (1:00am).  This can be done by setting
times for which raid eggs are known to be active.